feat(rebranding notification): move all logic to single file
