+++ .changeset/rich-experts-talk.md
@@ -3,4 +3,4 @@
 '@commercetools-frontend/assets': minor
 ---
 
-Users can now see a dialog on notifying them of the new rebranding work. When the information modal is closed, the message should not be seen again when the user comes back to the site.
+Included a new temporary dialog to notify users about the new colour schema.
